// backend/src/models/Titulo.ts
export class Titulo{
    id: number = 0;
    nome: string = '';
    descricao: string = '';
    data_lancamento: string = '';
    idioma: string = '';
    nota: number = 0;
    generos: string[] = [];
    produtoras: string[] = [];
    poster: string = 'https://image.tmdb.org/t/p/original';
    homepage: string = '';
    imdb_id: string = '';
}